+++ client/src/app/layout/ui/Navbar.tsx
@@ -1,6 +1,6 @@
 import { LEVEL_MAP, navItems } from '@/app/layout/data/navItems';
-import { useProfileQuery } from '@/features/auth/hooks/useProfileQuery';
 import { AuthButton } from '@/features/auth/ui/AuthButton';
+import { useProfileQuery } from '@/features/profile/api/useProfileQuery';
 import { useOutsideClick } from '@/shared/hooks/useOutsideClick';
 import { useToggle } from '@/shared/hooks/useToggle';
 import { sendEvent } from '@/shared/lib/ga';
@@ -10,10 +10,10 @@ import { NavigatorsBar } from '@/widgets/navigatorsBar';
 import { useCheckIfLoggedInQuery } from '@/features/auth/hooks/useCheckIfLoggedInQuery';
 import { useNotificationsQuery } from '@/features/notification/hooks/useNotificationsQuery';
 import { HomePageAnalyticsEvent } from '@/shared/lib/ga/analyticsEvent';
+import { EXPBar } from '@/widgets/EXPBar/EXPBar';
 import { useRef } from 'react';
 import { Link, useLocation } from 'react-router';
 import * as S from './Navbar.styles';
-import { EXPBar } from '@/widgets/EXPBar/EXPBar';
 
 export type Level = 'METEOR' | 'ASTEROID' | 'COMET';
