+++ client/src/features/auth/hooks/useLoginMutation.ts
@@ -1,6 +1,6 @@
 import { queryClient } from '@/app/lib/queryClient';
-import { getProfile } from '@/features/auth/api/getProfile';
 import { loginUser } from '@/features/auth/api/loginUser';
+import { getProfile } from '@/features/profile/api/useProfileQuery';
 import { useToast } from '@/shared/hooks/useToast';
 import { useMutation } from '@tanstack/react-query';
 import { useNavigate } from 'react-router';
