--- client/src/features/my/api/getProfile.ts
@@ -1,6 +1,0 @@
-import { api } from '@/app/lib/api';
-
-export const getProfile = async () => {
-  const response = await api.get('/me/profile');
-  return response.data;
-};
