+++ client/src/features/profile/types/profile.ts
@@ -1,8 +1,11 @@
 export interface Profile {
   nickname: string;
+  email: string;
+  availableStar: number;
   level: string;
   expStar: number;
   nextStepExp: number;
+  loginType: string;
 }
 
 export interface ProfileResponse {
