+++ client/src/features/profile/ui/RewardHistoryPagination.styles.ts
