+++ client/src/features/profile/ui/RewardHistoryTable.styles.ts
